test: Use cookie as token transfer method for existing tests
