+++ examples/with-fastapi/with-thirdpartyemailpassword/main.py
@@ -55,7 +55,7 @@ init(
     ),
     framework="fastapi",
     recipe_list=[
-        session.init(),
+        session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         emailverification.init("REQUIRED"),
         thirdpartyemailpassword.init(
             providers=[
