+++ examples/with-flask/with-thirdpartyemailpassword/app.py
@@ -48,7 +48,7 @@ init(
     ),
     framework="flask",
     recipe_list=[
-        session.init(),
+        session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         emailverification.init("REQUIRED"),
         thirdpartyemailpassword.init(
             providers=[
