+++ supertokens_python/recipe/session/utils.py
@@ -29,7 +29,6 @@ from supertokens_python.utils import (
     send_non_200_response,
     send_non_200_response_with_message,
 )
-from tldextract import extract  # type: ignore
 from typing_extensions import Literal
 
 from ...types import MaybeAwaitable
@@ -105,24 +104,6 @@ def get_url_scheme(url: str) -> str:
     return url_obj.scheme
 
 
-def get_top_level_domain_for_same_site_resolution(url: str) -> str:
-    url_obj = urlparse(url)
-    hostname = url_obj.hostname
-
-    if hostname is None:
-        raise Exception("Should not come here")
-
-    if hostname.startswith("localhost") or is_an_ip_address(hostname):
-        return "localhost"
-    parsed_url: Any = extract(hostname, include_psl_private_domains=True)
-    if parsed_url.domain == "":  # type: ignore
-        raise Exception(
-            "Please make sure that the apiDomain and websiteDomain have correct values"
-        )
-
-    return parsed_url.domain + "." + parsed_url.suffix  # type: ignore
-
-
 class ErrorHandlers:
     def __init__(
         self,
@@ -157,7 +138,7 @@ class ErrorHandlers:
         response: BaseResponse,
     ) -> BaseResponse:
         log_debug_message("Clearing tokens because of TOKEN_THEFT_DETECTED response")
-        clear_session_from_all_token_transfer_methods(recipe, request, response)
+        clear_session_from_all_token_transfer_methods(response, recipe, request)
         return await resolve(
             self.__on_token_theft_detected(request, session_handle, user_id, response)
         )
@@ -178,7 +159,7 @@ class ErrorHandlers:
     ):
         if do_clear_cookies:
             log_debug_message("Clearing tokens because of UNAUTHORISED response")
-            clear_session_from_all_token_transfer_methods(recipe, request, response)
+            clear_session_from_all_token_transfer_methods(response, recipe, request)
         return await resolve(self.__on_unauthorised(request, message, response))
 
     async def on_invalid_claim(
@@ -434,12 +415,8 @@ def validate_and_normalise_user_input(
     cookie_domain = (
         normalise_session_scope(cookie_domain) if cookie_domain is not None else None
     )
-    top_level_api_domain = get_top_level_domain_for_same_site_resolution(
-        app_info.api_domain.get_as_string_dangerous()
-    )
-    top_level_website_domain = get_top_level_domain_for_same_site_resolution(
-        app_info.website_domain.get_as_string_dangerous()
-    )
+    top_level_api_domain = app_info.top_level_api_domain
+    top_level_website_domain = app_info.top_level_website_domain
 
     api_domain_scheme = get_url_scheme(app_info.api_domain.get_as_string_dangerous())
     website_domain_scheme = get_url_scheme(
