+++ tests/emailpassword/test_emaildelivery.py
@@ -109,7 +109,10 @@ async def test_reset_password_default_backward_compatibility(
             api_base_path="/auth",
         ),
         framework="fastapi",
-        recipe_list=[emailpassword.init(), session.init()],
+        recipe_list=[
+            emailpassword.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
     )
     start_st()
 
@@ -157,7 +160,10 @@ async def test_reset_password_default_backward_compatibility_suppress_error(
             api_base_path="/auth",
         ),
         framework="fastapi",
-        recipe_list=[emailpassword.init(), session.init()],
+        recipe_list=[
+            emailpassword.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
     )
     start_st()
 
@@ -217,7 +223,7 @@ async def test_reset_password_backward_compatibility(driver_config_client: TestC
                     create_and_send_custom_email=custom_create_and_send_custom_email,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -269,7 +275,7 @@ async def test_reset_password_custom_override(driver_config_client: TestClient):
                     override=email_delivery_override,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -386,7 +392,7 @@ async def test_reset_password_smtp_service(driver_config_client: TestClient):
                     override=email_delivery_override,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -488,7 +494,7 @@ async def test_reset_password_for_non_existent_user(driver_config_client: TestCl
                     override=email_delivery_override,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -529,7 +535,7 @@ async def test_email_verification_default_backward_compatibility(
         recipe_list=[
             emailverification.init(mode="OPTIONAL"),
             emailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -594,7 +600,7 @@ async def test_email_verification_default_backward_compatibility_suppress_error(
         recipe_list=[
             emailverification.init(mode="OPTIONAL"),
             emailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -669,7 +675,7 @@ async def test_email_verification_backward_compatibility(
                 create_and_send_custom_email=custom_create_and_send_custom_email,
             ),
             emailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -739,7 +745,7 @@ async def test_email_verification_custom_override(driver_config_client: TestClie
                 ),
             ),
             emailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -870,7 +876,7 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
                 ),
             ),
             emailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
