+++ tests/passwordless/test_smsdelivery.py
@@ -108,7 +108,7 @@ async def test_pless_login_default_backward_compatibility(
                 contact_config=passwordless.ContactPhoneOnlyConfig(),
                 flow_type="USER_INPUT_CODE_AND_MAGIC_LINK",
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -204,7 +204,7 @@ async def test_pless_login_default_backward_compatibility_no_suppress_error_for_
                 contact_config=passwordless.ContactPhoneOnlyConfig(),
                 flow_type="USER_INPUT_CODE_AND_MAGIC_LINK",
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -268,7 +268,7 @@ async def test_pless_login_default_backward_compatibility_suppress_error_for_429
                 contact_config=passwordless.ContactPhoneOnlyConfig(),
                 flow_type="USER_INPUT_CODE_AND_MAGIC_LINK",
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -345,7 +345,7 @@ async def test_pless_login_backward_compatibility(driver_config_client: TestClie
                 ),
                 flow_type="USER_INPUT_CODE_AND_MAGIC_LINK",
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -425,7 +425,7 @@ async def test_pless_login_custom_override(driver_config_client: TestClient):
                     override=sms_delivery_override,
                 ),
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -573,7 +573,7 @@ async def test_pless_login_twilio_service(driver_config_client: TestClient):
                     override=sms_delivery_override,
                 ),
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
