+++ tests/sessions/claims/test_validate_claims_for_session_handle.py
@@ -49,7 +49,12 @@ async def test_should_return_the_right_validation_errors():
 
 
 async def test_should_work_for_not_existing_handle():
-    new_st_init = {**st_init_common_args, "recipe_list": [session.init()]}
+    new_st_init = {
+        **st_init_common_args,
+        "recipe_list": [
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie")
+        ],
+    }
     init(**new_st_init)  # type: ignore
     start_st()
