+++ tests/test_middleware.py
@@ -52,7 +52,10 @@ async def test_rid_with_session_and_non_existent_api_in_session_recipe_gives_404
             website_domain="supertokens.io",
         ),
         framework="fastapi",
-        recipe_list=[session.init(), emailpassword.init()],
+        recipe_list=[
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+            emailpassword.init(),
+        ],
     )
     start_st()
 
@@ -72,7 +75,10 @@ async def test_no_rid_with_existent_API_does_not_give_404(
             website_domain="supertokens.io",
         ),
         framework="fastapi",
-        recipe_list=[session.init(), emailpassword.init()],
+        recipe_list=[
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+            emailpassword.init(),
+        ],
     )
     start_st()
 
@@ -92,7 +98,10 @@ async def test_rid_as_anticsrf_with_existent_API_does_not_give_404(
             website_domain="supertokens.io",
         ),
         framework="fastapi",
-        recipe_list=[session.init(), emailpassword.init()],
+        recipe_list=[
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+            emailpassword.init(),
+        ],
     )
     start_st()
 
@@ -114,7 +123,10 @@ async def test_random_rid_with_existent_API_does_gives_404(
             website_domain="supertokens.io",
         ),
         framework="fastapi",
-        recipe_list=[session.init(), emailpassword.init()],
+        recipe_list=[
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+            emailpassword.init(),
+        ],
     )
     start_st()
