+++ tests/thirdpartyemailpassword/test_email_delivery.py
@@ -119,7 +119,10 @@ async def test_reset_password_default_backward_compatibility(
             api_base_path="/auth",
         ),
         framework="fastapi",
-        recipe_list=[thirdpartyemailpassword.init(), session.init()],
+        recipe_list=[
+            thirdpartyemailpassword.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
     )
     start_st()
 
@@ -168,7 +171,10 @@ async def test_reset_password_default_backward_compatibility_suppress_error(
             api_base_path="/auth",
         ),
         framework="fastapi",
-        recipe_list=[thirdpartyemailpassword.init(), session.init()],
+        recipe_list=[
+            thirdpartyemailpassword.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
     )
     start_st()
 
@@ -229,7 +235,7 @@ async def test_reset_password_backward_compatibility(driver_config_client: TestC
                     create_and_send_custom_email=custom_create_and_send_custom_email,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -281,7 +287,7 @@ async def test_reset_password_custom_override(driver_config_client: TestClient):
                     override=email_delivery_override,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -398,7 +404,7 @@ async def test_reset_password_smtp_service(driver_config_client: TestClient):
                     override=email_delivery_override,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -444,7 +450,7 @@ async def test_reset_password_backward_compatibility_non_existent_user(
                     create_and_send_custom_email=custom_create_and_send_custom_email,
                 )
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -481,7 +487,7 @@ async def test_email_verification_default_backward_compatibility(
         recipe_list=[
             emailverification.init(mode="OPTIONAL"),
             thirdpartyemailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -546,7 +552,7 @@ async def test_email_verification_default_backward_compatibility_suppress_error(
         recipe_list=[
             emailverification.init(mode="OPTIONAL"),
             thirdpartyemailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -621,7 +627,7 @@ async def test_email_verification_backward_compatibility(
                 create_and_send_custom_email=custom_create_and_send_custom_email,
             ),
             thirdpartyemailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -691,7 +697,7 @@ async def test_email_verification_custom_override(driver_config_client: TestClie
                 ),
             ),
             thirdpartyemailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -822,7 +828,7 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
                 ),
             ),
             thirdpartyemailpassword.init(),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -886,7 +892,7 @@ async def test_reset_password_backward_compatibility_thirdparty_user(
                     create_and_send_custom_email=custom_create_and_send_custom_email,
                 ),
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -950,7 +956,7 @@ async def test_email_verification_backward_compatibility_thirdparty_user(
                     Github(client_id="", client_secret="")
                 ],  # Note: Provider must be passed to init TP recipe
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
