+++ tests/thirdpartypasswordless/test_emaildelivery.py
@@ -140,7 +140,7 @@ async def test_email_verify_default_backward_compatibility(
                     Github(client_id="", client_secret="")
                 ],  # Note: providers must be set to init tp recipe
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -222,7 +222,7 @@ async def test_email_verify_backward_compatibility(driver_config_client: TestCli
                     Github(client_id="", client_secret="")
                 ],  # Note: providers must be set to init tp recipe
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -301,7 +301,7 @@ async def test_email_verify_custom_override(driver_config_client: TestClient):
                     Github(client_id="", client_secret="")
                 ],  # Note: providers must be set to init tp recipe
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -442,7 +442,7 @@ async def test_email_verify_smtp_service(driver_config_client: TestClient):
                     Github(client_id="", client_secret="")
                 ],  # Note: providers must be set to init tp recipe
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -557,7 +557,7 @@ async def test_email_verify_for_pless_user_no_callback():
                     override=email_delivery_override,
                 ),
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -603,7 +603,7 @@ async def test_pless_login_default_backward_compatibility(
                 contact_config=passwordless.ContactEmailOnlyConfig(),
                 flow_type="USER_INPUT_CODE_AND_MAGIC_LINK",
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -693,7 +693,7 @@ async def test_pless_login_default_backward_compatibility_no_suppress_error(
                 contact_config=passwordless.ContactEmailOnlyConfig(),
                 flow_type="USER_INPUT_CODE_AND_MAGIC_LINK",
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -765,7 +765,7 @@ async def test_pless_login_backward_compatibility(driver_config_client: TestClie
                 ),
                 flow_type="USER_INPUT_CODE_AND_MAGIC_LINK",
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -847,7 +847,7 @@ async def test_pless_login_custom_override(driver_config_client: TestClient):
                     override=email_delivery_override,
                 ),
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
@@ -996,7 +996,7 @@ async def test_pless_login_smtp_service(driver_config_client: TestClient):
                     override=email_delivery_override,
                 ),
             ),
-            session.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
         ],
     )
     start_st()
