+++ tests/userroles/test_get_roles_for_user.py
@@ -43,7 +43,10 @@ async def test_get_roles_for_user():
             website_domain="supertokens.io",
         ),
         framework="fastapi",
-        recipe_list=[userroles.init(), session.init()],
+        recipe_list=[
+            userroles.init(),
+            session.init(get_token_transfer_method=lambda _, __, ___: "cookie"),
+        ],
     )
     start_st()
